refactor: abstract PropertyInspector class to separate file
