+++ src/plugin/ui/__tests__/routing.test.ts
@@ -1,15 +1,16 @@
-import { Action, JsonValue, MessageRequest, MessageRequestOptions } from "../..";
+import { Action, MessageRequest, type MessageRequestOptions } from "../..";
 import type { DidReceivePropertyInspectorMessage, SendToPropertyInspector } from "../../../api";
 import type { RawMessageRequest } from "../../../common/messaging/message";
 import { MessageResponder } from "../../../common/messaging/responder";
 import { connection } from "../../connection";
-import { PropertyInspector, getCurrentUI, router } from "../routing";
+import { PropertyInspector } from "../property-inspector";
+import { getCurrentUI, router } from "../routing";
 
 jest.mock("../../connection");
 jest.mock("../../logging");
 jest.mock("../../manifest");
 
-describe("getCurrentUI", () => {
+describe("current UI", () => {
 	/**
 	 * Asserts {@link getCurrentUI} is set when the connection emits `propertyInspectorDidAppear`.
 	 */
@@ -22,6 +23,7 @@ describe("getCurrentUI", () => {
 			event: "propertyInspectorDidAppear"
 		});
 
+		// Act.
 		const current = getCurrentUI();
 
 		// Assert.
@@ -58,101 +60,33 @@ describe("getCurrentUI", () => {
 		// Assert.
 		expect(current).toBeUndefined();
 	});
-});
-
-describe("PropertyInspector", () => {
-	/**
-	 * Asserts {@link PropertyInspector} initializes it's context on construction.
-	 */
-	it("initializes context", () => {
-		// Arrange, act.
-		const pi = new PropertyInspector({
-			action: "com.elgato.test.one",
-			context: "abc123",
-			device: "dev123"
-		});
-
-		// Assert.
-		expect(pi.deviceId).toBe("dev123");
-		expect(pi.id).toBe("abc123");
-		expect(pi.manifestId).toBe("com.elgato.test.one");
-	});
-
-	describe("fetch", () => {
-		/**
-		 * Asserts {@link PropertyInspector.fetch} forwards the path/body to {@link router.fetch}.
-		 */
-		test("path and body", async () => {
-			// Arrange.
-			const spyOnFetch = jest.spyOn(router, "fetch");
-			const pi = new PropertyInspector({
-				action: "com.elgato.test.one",
-				context: "abc123",
-				device: "dev123"
-			});
-
-			// Act.
-			await pi.fetch("/hello", { name: "Elgato" });
-
-			// Assert.
-			expect(spyOnFetch).toBeCalledTimes(1);
-			expect(spyOnFetch).toHaveBeenLastCalledWith<[string, JsonValue]>("/hello", { name: "Elgato" });
-		});
-
-		/**
-		 * Asserts {@link PropertyInspector.fetch} forwards the request {@link router.fetch}.
-		 */
-		test("request", async () => {
-			// Arrange.
-			const spyOnFetch = jest.spyOn(router, "fetch");
-			const pi = new PropertyInspector({
-				action: "com.elgato.test.one",
-				context: "abc123",
-				device: "dev123"
-			});
-
-			// Act.
-			await pi.fetch({
-				path: "/hello",
-				body: { name: "Elgato" },
-				timeout: 1000,
-				unidirectional: true
-			});
-
-			// Assert.
-			expect(spyOnFetch).toBeCalledTimes(1);
-			expect(spyOnFetch).toHaveBeenLastCalledWith<[MessageRequestOptions]>({
-				path: "/hello",
-				body: { name: "Elgato" },
-				timeout: 1000,
-				unidirectional: true
-			});
-		});
-	});
 
 	/**
-	 * Asserts {@link PropertyInspector.sendMessage} sends the event to the {@link connection}.
+	 * Asserts {@link getCurrentUI} uses the {@link router}.
 	 */
-	it("sends messages", async () => {
+	it("proxies fetch to router", async () => {
 		// Arrange.
-		const spyOnSend = jest.spyOn(connection, "send");
-		const pi = new PropertyInspector({
+		const spyOnFetch = jest.spyOn(router, "fetch");
+		connection.emit("propertyInspectorDidAppear", {
 			action: "com.elgato.test.one",
 			context: "abc123",
-			device: "dev123"
+			device: "dev123",
+			event: "propertyInspectorDidAppear"
 		});
 
 		// Act.
-		await pi.sendMessage({ message: "Hello world" });
+		await getCurrentUI()!.fetch({
+			path: "/test",
+			unidirectional: true,
+			timeout: 1
+		});
 
 		// Assert.
-		expect(spyOnSend).toBeCalledTimes(1);
-		expect(spyOnSend).toHaveBeenLastCalledWith<[SendToPropertyInspector]>({
-			context: "abc123",
-			event: "sendToPropertyInspector",
-			payload: {
-				message: "Hello world"
-			}
+		expect(spyOnFetch).toBeCalledTimes(1);
+		expect(spyOnFetch).toHaveBeenCalledWith<[MessageRequestOptions]>({
+			path: "/test",
+			timeout: 1,
+			unidirectional: true
 		});
 	});
 });
